Move remaining `str_*` functions to `base/str.cpp/h`
Also move the `#include <cinttypes>` and the string format constants `PRIzu` etc. to `str.h` as it seems like the most appropriate location for them.
