+++ src/base/hash.cpp
@@ -1,7 +1,8 @@
 #include "hash.h"
 
 #include "hash_ctxt.h"
-#include "system.h"
+#include "mem.h"
+#include "str.h"
 
 static void digest_str(const unsigned char *digest, size_t digest_len, char *str, size_t max_len)
 {
