+++ src/base/unicode/confusables.cpp
@@ -1,6 +1,6 @@
 #include "confusables.h"
 
-#include <base/system.h>
+#include <base/str.h>
 
 #include <cstddef>
