+++ src/engine/server/databases/connection.cpp
@@ -1,6 +1,6 @@
 #include "connection.h"
 
-#include <base/system.h> // TODO: replace with str.h after moving str_format
+#include <base/str.h>
 
 IDbConnection::IDbConnection(const char *pPrefix)
 {
